+++ packages/migration/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @larose-ui/migration
 
+## Unreleased
+
+### Minor Changes
+
+- Add smart component/module contribution scaffold plans used by `larose contribute`.
+
 ## 0.2.0
 
 ### Minor Changes